+++ python/sedonadb/python/sedonadb/functions/aggregate.py
@@ -1,0 +1,32 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+
+from sedonadb.expr.expression import AggregateUdf
+
+
+class AggregateFunctions:
+    def __init__(self, ctx):
+        self._ctx = ctx
+
+    def __getitem__(self, key: str) -> AggregateUdf:
+        return AggregateUdf(self._ctx._impl.aggregate_udf(key))
+
+    def __getattr__(self, name) -> AggregateUdf:
+        return AggregateUdf(self._ctx._impl.aggregate_udf(name))
+
+    def __dir__(self) -> str:
+        return self._ctx._impl.list_aggregate_udfs()
